Refine dashboard layout and headings
