+++ presentation/frontend/src/components/ChartSelector.vue
